move proxy stuff from EntitySystem to EntManProxy
